requirements.txt
python-dateutil==2.8.2
python-liquid==1.10.2
six==1.16.0
typing_extensions==4.9.0
